requirements.txt
jupyter
pandas
numpy
matplotlib
psycopg2-binary